built page 1 and moved the button control dealy thingy
i probably should've checked to see if slick.js would've made the pages responsive, but then I was like "nah I wanna used grid and make it fit myself" so I spent like 8 hours doing that and then trying to figure out which classes control what, and what the code does. i'm not entirely sure but i'm going to get the menu to correspond with the buttons.